+++ - Test/Audio_File_Tester/Audio_File_Tester/main.cpp
@@ -9,9 +9,9 @@
 #include <chrono>
 #include <limits>
 
-#include "../../../include/AudioFile/in_file.hpp"
-#include "../../../include/AudioFile/out_file.hpp"
-#include "../../../include/AudioFile/utilities.hpp"
+#include "../../../include/audio_file/in_file.hpp"
+#include "../../../include/audio_file/out_file.hpp"
+#include "../../../include/audio_file/utilities.hpp"
 #include "../../../include/random_generator.hpp"
