+++ - Test/Window_Tester/Window_Tester/main.cpp
@@ -8,7 +8,7 @@
 #include <random>
 
 #include "../../../include/window_functions.hpp"
-#include "../../../include/AudioFile/out_file.hpp"
+#include "../../../include/audio_file/out_file.hpp"
 
 // Output
