+++ include/audio_file/extended_double.hpp
